dom/tests/mochitest/bugs/test_bug404748.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=404748
-->
<head>
  <title>Test for Bug 404748</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=404748">Mozilla Bug 404748</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 404748 **/
var setterCalled = false;
HTMLDivElement.prototype.
  __defineSetter__("onclick", function () { window.setterCalled = true });
$("content").onclick = function () {};

is(setterCalled, true, "Setter should have been called");
</script>
</pre>
</body>
</html>
